Refactoring the EBM code

+++ scripts/test_EBM_clip_vector.py
@@ -1258,7 +1258,8 @@ def get_clip_embeddings_by_tag_with_data_augmentation(id_classes,label_value):
     class_A_clips = []
     for image in class_A_images:
         clip_emb = image_embedder.get_image_features(image)
-        class_A_clips.append(clip_emb.float())
+        clip_emb = clip_emb.float()
+        class_A_clips.append(clip_emb)
 
 
     # Create labels
